test(Changex.Log): add missing test for passing in a directory
`Changex.Log.log` takes an optional parameter for the directory to
perform a log on. If this directory is specified then a `--git-dir`
argument will be prepended to the args - this will perform git log on
the specified directory.

The way this argument is prepended has been refectored to ensure the git
log format is only specified once.