feat(Mix.Tasks.Changex.Diff): pass a dir through as a --dir argument
Previously the directory was passed to the mix task as the first
argument. In future this will be reserved for the start date of for
the log. It can now be passed as `--dir /path/to/dir`

BREAKING CHANGE: calling `mix changex.diff /path/to/dir` will no longer
work. `mix changex.diff --dir /path/to/dir` is now required.