feat(Changer.Formatter.Terminal): print version with changelog
When printing a changelog, the version is printed as a heading. The
types are now printed as sub headings. The version will default to the
current version defined in mix. If a `last` commit identifier is passed
then that is used as the version. This is used if you are retroactively
generating a changelog for a project.