feat(Changex.Formatter.Markdown): return a string instead of IO.puts
Previously this function would output the string as markdown to the
terminal using `IO.puts` - this technique was not ideal for two reasons

 1. It does not lend itself well to testing
 2. It will only put to the terminal and not to some other IO such as
    a file.

So that the output is still presented to the terminal in the mix task,
`IO.puts` is now called inside the mix task.

The private functions inside the module have been renamed to `build`
instead of `output` as that makes more semantic sense. The public
function is still called `output` to maintain consistency with the
terminal formatter.

A test has been added for this formatter.