fix(Mix.Tasks.Changex.Diff): use correct dir for first git ref
When a `--dir` flag is specified, previously it was using the latest tag
for the current directory. This would mean that if the repository you
wish to generate a changelog for does not have a tag matching the
repository for the current directory then an error would be raised.

The tag is now looked up by the directory passed in as an argument.