fix(Changex.Formatter.Elixir): remove new lines for empty types
Previously if there was an empty type in the commits object (for example
no bug fixes) then a new line character would be output for that type.
This would result in an additional space between the version heading and
the first type. For example:

  ## v0.0.1

  --extra whitespace line--
  * Bug fixes
    * [policy] remove reference to data retention length
  ...

This now displays as:

  ## v0.0.1

  * Bug fixes
    * [policy] remove reference to data retention length
  ...

A new test is in place to ensure this is formatter correctly. It is now
required in the tests to pass the strings through `String.rstrip` to
ensure there is no newline at the end of the expected output.