refactor(Mix.Tasks.Changex.Update): separate format builder to own def
The responsibility of building the changelog is now passed to a new
`build` function within the mix task. This function takes the commits
nested object, the previous contents of the changelog and a format. It
then generates the changes for the repository and prepends them to the
previous changelog file.

This change will allow additional formatters to be added easily.