feat(Changex.Formatter.Elixir): pass options list instead of version
The version argument that is passed through to the elixir parser is now
a keyword list. This will easily enable other options to be passed in
the future (such as the GitHub url)

The other formatter will be changed to a similar format in the future.

BREAKING CHANGE: Calling `Changex.Formatter.Elixir.format` with a string
as the second argument is no longer valid. Instead you can call
`Changex.Formatter.Elixir.format` with an options list as the second
argument. This list can contain a `version` key.