feat(Changex.Formatter): create a module to be used by other formatters
A new module `Changex.Formatter` has been created to be used by modules
that wish to be used as a formatter (such as the built in
`Changex.Formatter.Markdown` or `Changex.Formatter.Terminal` as well as
user defined modules)

These can be used in the following way:

    defmodule MyCustomFormatter do
      use Changex.Formatter
    end

Currently this only imports an internal `Utils` function however this
can easily be extended in the future to import many other modules. The
Utils module specifies a `current_version` function - this returns the
version specified in the mix config prepended by the letter "v". This
ensures consistency between the version specified in `mix.exs` (such as
0.1.0) and git tags (such as v0.0.9)