fix(Changex.Formatter.Markdown): use Changex.Formatter.current_version
Previously, when running `mix changex.diff` the output would use version
specified in `mix.exs` - this would result in inconsistent results when
comparing with git tags. The `current_version` function that is included
when `use Changex.Formatter` is called is now used instead. The test has
also been updated to reflect this change.