feat(Changex.Formatter.Markdown): pass version in opts instead arguments
Previously there was an inconsistency between
`Changex.Formatter.Markdown` and `Changex.Formatter.Elixir` where in
`Elixir` it would take the version as part of the opts list and in
`Markdown` it was a separate argument. In commit
c3919d4f60047cc80af7b3ce7ac87be5826500d4 it is made clear that the
intention was to pass an opts instead of the version, however this
change was node made for `Markdown` at the same time. The format
function between the two modules is now consistent.

The `update` and `diff` tasks have been updated to use the new function
with an arity of 2.

BREAKING CHANGE: Calling `Changex.Formatter.Markdown.format` with an
arity of 3 is no longer valid. The `version` should be specified as a
keyword in the second argument which is a keyword list.